trade-service 최근내역 조회 지수표기 됌

+++ trading-service/trading-application/src/main/java/shop/shportfolio/trading/application/command/track/response/TradeTickTrackResponse.java
@@ -3,6 +3,8 @@ package shop.shportfolio.trading.application.command.track.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class TradeTickTrackResponse {
@@ -20,16 +22,16 @@ public class TradeTickTrackResponse {
     private final Long timestamp;
 
     /** 체결 가격 */
-    private final Double tradePrice;
+    private final BigDecimal tradePrice;
 
     /** 체결량 */
-    private final Double tradeVolume;
+    private final BigDecimal tradeVolume;
 
     /** 전일 종가(UTC 0시 기준) */
-    private final Double prevClosingPrice;
+    private final BigDecimal prevClosingPrice;
 
     /** 변화량 */
-    private final Double changePrice;
+    private final BigDecimal changePrice;
 
     /** 매도/매수 */
     private final String askBid;
